feature: added DP8, animation fix
