+++ data/input.txt
@@ -1,6 +1,6 @@
-0 1 0
-0.01 0 0.01
 0.0 0.0 0.0
-0.0 0.0 300.0
+0.0 0 0.0628
+0.0 0.0 0.0
+10 0.0 0.0
 1 1
-2000 0.01 1000
+100 0.01 10
