+++ data/methods/38_rule.txt
@@ -1,0 +1,5 @@
+4
+0.333333333
+-0.333333333 1
+1 -1 1
+0.125 0.375 0.375 0.125
