--- data/methods/RK4_classic.txt
@@ -1,5 +1,0 @@
-4
-0.5
-0.0 0.5
-0.0 0.0 1.0
-0.166666667 0.333333333 0.333333333 0.166666667
