+++ data/methods/dp5.txt
@@ -1,0 +1,8 @@
+7
+0.2
+0.075 0.225
+0.977777778 -3.733333333 3.555555556
+2.952598689 -11.595793324 9.822892852 -0.290809328
+2.846275253 -10.757575758 8.906422718 0.278409091 -0.273531304
+0.091145833 0 0.449236298 0.651041667 -0.322376179 0.130952381
+0.091145833 0 0.449236298 0.651041667 -0.322376179 0.130952381 0
