+++ data/methods/dp8.txt
@@ -1,0 +1,15 @@
+14
+0.055555556
+0.020833333 0.0625
+0.03125 0 0.09375
+0.3125 0 -1.171875 1.171875
+0.0375 0 0 0.1875 0.15
+0.047910137 0 0 0.112248713 -0.025505674 0.012846824
+0.01691799 0 0 0.387848278 0.03597737 0.196970214 -0.172713852
+0.069095753 0 0 -0.634247977 -0.161197575 0.138650309 0.940928614 0.211636326
+0.183556997 0 0 -2.468768084 -0.291286888 -0.02647302 2.847838764 0.281387331 0.1237449
+-1.215424817 0 0 16.672608666 0.915741828 -6.056605804 -16.003573594 14.849303086 -13.371575735 5.134182648
+0.258860916 0 0 -4.774485785 -0.435093014 -3.049483332 5.57792004 6.15583159 -5.062104587 2.193926173 0.134627999
+0.8224276 0 0 -11.658673257 -0.757622117 0.713973588 12.075774987 -2.127659114 1.990166207 -0.234286472 0.175898578 0
+0.041747491 0 0 0 0 -0.055452329 0.239312807 0.703510669 -0.759759614 0.660563031 0.158187483 -0.238109539 0.25
+0.041747491 0 0 0 0 -0.055452329 0.239312807 0.703510669 -0.759759614 0.660563031 0.158187483 -0.238109539 0.25 0
