+++ data/methods/fehlberg5.txt
@@ -1,0 +1,7 @@
+6
+0.25
+0.09375 0.28125
+0.879380974 -3.277196177 3.320892126
+2.032407407 -8 7.173489279 -0.205896686
+-0.296296296 2 -1.381676413 0.45297271 -0.275
+0.118518519 0 0.518986355 0.50613149 -0.18 0.036363636
