+++ data/methods/ralstons.txt
@@ -1,0 +1,5 @@
+4
+0.4
+0.29697761 0.15875964
+0.21810040 -3.05096516 3.83286476
+0.17476028 -0.55148066 1.20553560 0.17118478
