+++ data/methods/rk4_classic.txt
@@ -1,0 +1,5 @@
+4
+0.5
+0.0 0.5
+0.0 0.0 1.0
+0.166666667 0.333333333 0.333333333 0.166666667
