--- plot/.gitignore
@@ -1,2 +1,0 @@
-*~
-*.pyc
